Lightning-Lens/lightning_lens/requirements.txt
grpcio>=1.54.0
protobuf>=4.21.0
googleapis-common-protos>=1.56.0
PyYAML>=6.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
torch>=2.0.0
networkx>=2.8.0
plotly>=5.13.0
pytest>=7.3.1
pytest-mock>=3.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
websockets>=10.0
flask>=2.0.0